audio: fix WAVE audio corruption at end of sample
Main issue is that some RIFF WAV files have additional non-audio chunks at the end of the file, for meta data. We were mixing this into the audio buffer, causing noise.

+++ decoder.h
@@ -26,13 +26,13 @@ typedef struct
 typedef struct
 {
    OggVorbis_File vf;
-   vorbis_info *info;
+   vorbis_info*   info;
 } dec_OggData;
 
 typedef struct
 {
-   void* fp;
-   int   pos;
+   void*      fp;
+   intmax_t   pos;
    wavhead_t head;
 } dec_WavData;
